[AI] Extract findTsConfig into standalone file
