+++ packages/telefunc/node/shared/transformer/generateShield/findTsConfig.spec.ts
@@ -1,4 +1,4 @@
-import { findTsConfig } from './generateShield.js'
+import { findTsConfig } from './findTsConfig.js'
 import { describe, it, expect, afterEach } from 'vitest'
 import fs from 'node:fs'
 import path from 'node:path'
